docs: rearranged the menu items
The rearrangement is done with the approach that when the user lands at webiny.com/docs, the sequence of docs defines the Webiny learning path. Also, article searching and navigation should be easy for the user based on the Webiny application they are using.

- Learn the Webiny Overview
- Install and Play
- Learn Core concept
- Learn how to build custom applications with Webiny
- After this, the user can navigate to application-specific articles, as we are aware not everybody is using all the applications of Webiny.
If someone uses CMS, they don't look for Page Builder most of the time. So, it will be good to have all the application-specific articles in one place, e.g. Headless CMS heading has all the articles related to Headless CMS under it.
- All the articles related to Deployment are grouped in one place
- Key Topics are the advanced topic to learn in the Webiny journey
- Security & Performance is the next topics to explore
- Finally, we have the Release notes and upgrade guides.