Request to update full form of SAM and a query
Updated SAM to AWS Serverless Application Model (SAM), if it is indeed the correct reference. 
Reference: https://aws.amazon.com/serverless/sam/

Query - An email from Steve to Jeremy is included. It does not have context, or i might be missing it. We can add a leading sentence for the conversation, if feasible. For example, "The following solution/suggestion can be used to create temporary stacks when you create a serverless infrastructure:". 

On a personal front, thanks. I did not realise that 'serverless services often cannot be run on your local machine'.