Request to update links
The 'Workflows' and 'Cloud Infrastructure' links seem to be incorrect. 

Under Info, 'Workflows' links to the same page. So, I updated the link to the Workflows topic, shown below.
https://www.webiny.com/docs/key-topics/ci-cd/workflows/

Under Watch out for extra costs, 'Cloud Infrastructure' links to the same page. So, I updated the link to the correct topic, shown below.
https://www.webiny.com/docs/key-topics/cloud-infrastructure/introduction