Request to update links
In the TIP section, 'Prerendering Pages' links to an empty page. Is there a dedicated 'Prerendering Pages' link for Page Builder, or can the below link be used? One way or another, please provide the link, and I will update the correct one.
https://www.webiny.com/docs/key-topics/cloud-infrastructure/website/prerendering-pages

In the last line, the 'in our GitHub repository' link does not work. Please provide the correct link, and I will update it.