The npm page, @webiny-contrib, is blank
In Step 4, 'Raise the Pull Request', the '@webiny-contrib' link leads to a page (https://www.npmjs.com/package/@webiny-contrib) with no entries. I assume this will be active when new community plugins are added? If so, please ignore this comment. If the link has to be updated, please provide the correct link, and I will update it.