Update changelog.mdx
In the Note, the 'Amazon DynamoDB' link is barely visible. Is it possible to change the hyperlink color or the Note background color for hyperlink visibility? If you agree, I will work with you to change it. 
In the 'Page Builder > Prerendering' section, the 'api/code/graphql/src/index.ts' link does not work