Request to update link
In the 'New types.ts Declaration File In GraphQL API Application Code (#1839)' section, under TIP, the 'ContextPlugin' link does not work. If you provide the link, I can update it.
In the 'Allow Links To Be Rendered in the Menu Group (#1857)' section, in the Note, the 'Extend Admin Area scaffold ' link is barely visible, since the hyperlink and background colors are the same. Is it possible to update one of the colors? If so, the link will be visible. If you agree, please let me know.