docs: update private model documentation (#731)
During the installation wizard of Webiny loading the private model
does cause an error. The reason is that loading models requires i18n,
but when the installation wizard launches, the setup of i18n is not
done yet.

This minor documentation patch will update the example code of the
books app to show case how to check if the CMS is already installed.

Only if the CMS is installed, the apps ContextPlugin code will be run.

Helped-by: Pavel Denisjuk <pavel@webiny.com>

Co-authored-by: Jens Lauterbach <jens.lauterbach@groundfog.cloud>